refactor : DateManager 수정
